cpp: Fix warnings, enforcing them
Fix the necessary warnings so that we can enable -Werror and stop
compilation at warnings.

+++ cpp/sieve.cpp
@@ -30,7 +30,7 @@ Sieve::fill(int nsize) {
 
 bool
 Sieve::is_prime(int num) {
-  if (num >= primes.size())
+  if (num >= (int)primes.size())
     new_length(num);
 
   return primes[num];
